# connectin-api-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Connect'in API",
        "description": "Documentation Swagger de l'API Connect'in",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/posts/{post}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Lister les commentaires d'un post",
                "operationId": "00a7052dfabc2d7ba47f1f5d5d62f176",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID du post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des commentaires"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Post introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comments": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Créer un commentaire sur un post",
                "operationId": "a0c0613c248bc727d51299e2eb03d2e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "content"
                                ],
                                "properties": {
                                    "post_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Super post !",
                                        "maxLength": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire créé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comments/{comment}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Modifier son commentaire",
                "operationId": "e2a8b55069571b4702e0bcc41b5645df",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Commentaire modifié",
                                        "maxLength": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commentaire modifié"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Commentaire introuvable"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Supprimer son commentaire",
                "operationId": "d4474416e583aa55f654c13d908f0be7",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire supprimé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Commentaire introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comments/{comment}/image": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Ajouter / remplacer l'image d'un commentaire",
                "operationId": "29bd1b9166c4c3ff4b9f5202e0d18332",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID du commentaire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image mise à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrer le token sous la forme: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}